Updated - Revert back to that send logs into Google cloud logger, - Add labels option module=vatz
